+++ src/frontend/apps/impress/src/features/docs/doc-export/blocks-mapping/dividerDocx.tsx
@@ -1,0 +1,24 @@
+import { Paragraph } from 'docx';
+
+import { useCunninghamTheme } from '@/cunningham';
+
+import { DocsExporterDocx } from '../types';
+
+export const blockMappingDividerDocx: DocsExporterDocx['mappings']['blockMapping']['divider'] =
+  () => {
+    const { colorsTokens } = useCunninghamTheme.getState();
+
+    return new Paragraph({
+      spacing: {
+        before: 200,
+      },
+      border: {
+        top: {
+          color: colorsTokens()['greyscale-300'],
+          size: 1,
+          style: 'single',
+          space: 1,
+        },
+      },
+    });
+  };