+++ src/frontend/apps/impress/src/features/docs/doc-export/blocks-mapping/index.ts
@@ -1,3 +1,5 @@
+export * from './dividerDocx';
+export * from './dividerPDF';
 export * from './headingPDF';
 export * from './paragraphPDF';
 export * from './quoteDocx';