+++ src/frontend/apps/impress/src/features/docs/doc-export/mappingDocx.tsx
@@ -1,12 +1,16 @@
 import { docxDefaultSchemaMappings } from '@blocknote/xl-docx-exporter';
 
-import { blockMappingQuoteDocx } from './blocks-mapping/';
+import {
+  blockMappingDividerDocx,
+  blockMappingQuoteDocx,
+} from './blocks-mapping/';
 import { DocsExporterDocx } from './types';
 
 export const docxDocsSchemaMappings: DocsExporterDocx['mappings'] = {
   ...docxDefaultSchemaMappings,
   blockMapping: {
     ...docxDefaultSchemaMappings.blockMapping,
+    divider: blockMappingDividerDocx,
     quote: blockMappingQuoteDocx,
   },
 };